Add project reports to the rocket pages for uxo and vidar3
This puts the final submitted versions of the project report pdfs for
2017 and 2018 on the team website, and includes a link to them from
the rocket pages for UXO and VIDARIII. We've been discussing how much
we value sharing information about our rockets, and this PR is an
extension of those discussions.

I put the reports on their respective rocket pages because it was the
minimum effort required. I considered making a new page to host all of
the reports from all time, but that seemed like more effort than it
was worth. I also only added the two most recent ones, if anyone feels
that all the old reports should be put up, I can do that.