Update the sponsorship package to 2020 (#93)
Upload the PDF for the new sponsorship package, and also remove the old
ones. There's no reason that we need to keep old sponsorship packages
floating around on the website. Also updated sponsors.html to just grab
sponsorship_package instead of needing to refer to it by year, so in the
future we can just replace the file.