Fix CSS issue where long names overflow profile
Previously, the position of member and alumni titles had a minimum
distance from the top, causing long names to overflow out of the
bottom of the profile image.
Changed positioning to be relative to the bottom of the parent element
to force text to overflow upwards.
Also reorganized unnecessary title and year divs into header tags
instead, since the only real difference between them was font
sizing.