Awards resizing (#97)
* Fix wrapping issues when resizing awards page

Fixed div organization so that resizing the screen no longer breaks
each awards section. The default is now to fit two awards per line if
the screen is large enough, otherwise only fit one. If the screen width
is too small for one award, the rest can be scrolled to on the same
line (instead of wrapping over to the next line)